Fixes
Edited State 28 so the french line is reduced from 10 bunkers to the max 6 bunkers.

Changed another German Localization for the FT.

Edited Germany FT with a few minor fixes as well as changing a lot of spacing as it was freaking me out. Sorry the spacing syntax is burned into my brain and refuses to not fix it haha.